[BE] feat: logout 기능 구현 (#266)
* refactor :네이밍 수정

* feat :인증 controller분리 및 logout 기능 구현

* refactor :사용하지 않는 클래스 삭제

* refactor :개행 제거

* refactor :네이밍 수정

* refactor :네이밍 수정

* refactor :컨벤션 적용

* refactor :merge 충돌 해겨

+++ backend/src/main/java/org/donggle/backend/auth/JwtTokenProvider.java
@@ -20,8 +20,8 @@ public class JwtTokenProvider {
     private static final String MEMBER_ID_KEY = "memberId";
 
     private final SecretKey key;
-    private final long accessTokenValidityInMilliseconds;
-    private final long refreshTokenValidityInMilliseconds;
+    private final int accessTokenValidityInMilliseconds;
+    private final int refreshTokenValidityInMilliseconds;
 
     public JwtTokenProvider(@Value("${security.jwt.token.secret-key}") final String secretKey,
                             @Value("${security.jwt.token.access-token-expire-length}") final int accessTokenValidityInMilliseconds,
